rename the hello package into earthquakes and make change in related files
