+++ src/main/java/earthquakes/Application.java
@@ -1,4 +1,4 @@
-package hello;
+package earthquakes;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
