+++ src/main/java/earthquakes/geojson/Feature.java
@@ -1,4 +1,4 @@
-package hello.geojson;
+package earthquakes.geojson;
 
 public class Feature {
   public String type;
