+++ src/main/java/earthquakes/services/MembershipService.java
@@ -1,4 +1,4 @@
-package hello.services;
+package earthquakes.services;
 
 import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
